stored all home screen components as separate components. Add tabular home-screen layout. Updated some layout colors and formatting
