+++ frontend/src/app.css
@@ -4,7 +4,7 @@
 
 body {
     padding: 2rem;
-    max-width: 1500px;
+    /* max-width: 1500px; */
     margin: auto;
 }
