+++ frontend/src/homepageComponents/BriefIntro/+page.svelte
@@ -1,0 +1,25 @@
+<!-- Welcome text and intro wealth inequality video -->
+<section class="">
+	<div>
+		<p class="my-5 indent-10">
+			Welcome to a platform for learning how the world works, what can be done about it, and how you
+			can get involved. Check out the intro video to learn why we bothered creating this space.
+		</p>
+		<p class="indent-10">
+			If you're not sure where to start, check out the guided learning path for a structured
+			school-like learning experience, or ask some common questions in the questions section. Either
+			way, browse around and learn how we can change the world for the better.
+		</p>
+	</div>
+	<!-- svelte-ignore a11y-media-has-caption -->
+	<div class="w-full flex justify-center my-10">
+		<iframe
+			class="w-full aspect-video max-w-4xl rounded-lg"
+			src="https://www.youtube.com/embed/uWSxzjyMNpU"
+			title="YouTube video player"
+			frameborder="0"
+			allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
+			allowfullscreen
+		/>
+	</div>
+</section>
