+++ frontend/src/homepageComponents/InteractiveWealth/+page.svelte
@@ -1,0 +1,31 @@
+<!-- Interactive wealth inequality -->
+<section>
+	<hgroup>
+		<h2 class="text-3xl mb-4">Would taxing the richest handful of billionaires solve anything?</h2>
+		<h3 class="indent-10">
+			It's almost impossible to visualize a billion dollars, let alone several trillion dollars, the
+			net wealth of around a 2,000 or so billionaires. Scroll to the right in the below graphic to
+			better understand just how imbalanced our system is that it allows and encourages around 2,000
+			people to keep so much of the profits while so many of the worlds problems could be solved by
+			a fraction of it.
+			<br />
+			<br />
+			Just keep scrolling, trust us! Take about 5 minutes to keep scrolling.
+		</h3>
+	</hgroup>
+	<div class="relative">
+		<figure>
+			<iframe
+				src="https://mkorostoff.github.io/1-pixel-wealth/"
+				height="600"
+				class="w-full my-10 rounded-lg"
+				title="Wealth visualized"
+			/>
+			<figcaption class="absolute right-5 bottom-10">
+				<a href="https://mkorostoff.github.io/1-pixel-wealth/"
+					>https://mkorostoff.github.io/1-pixel-wealth</a
+				>
+			</figcaption>
+		</figure>
+	</div>
+</section>
