+++ frontend/src/routes/+page.svelte
@@ -1,426 +1,25 @@
 <script lang="ts">
-	import { flip } from 'svelte/animate';
-	import { fade } from 'svelte/transition';
-	import FlipCard from '../components/FlipCard.svelte';
-
-	let cardStates = false;
-
-	function clicked() {
-		cardStates = !cardStates;
-		console.log(cardStates);
-	}
-
-	let cards = [
-		{
-			id: 1,
-			question: 'this is question 1?',
-			answer: 'This is the answer for question 1'
-		},
-		{
-			id: 2,
-			question: 'this is question 2?',
-			answer: 'This is the answer for question 2'
-		},
-		{
-			id: 3,
-			question: 'this is question 3?',
-			answer: 'This is the answer for question 3'
-		}
-	];
+	import Tabs from '../components/Tabs.svelte';
+	import Categories from '../homepageComponents/Categories/+page.svelte';
 </script>
 
-<header class="lg:w-8/12 flex flex-col m-auto" transition:fade>
+<header class="lg:w-8/12 flex flex-col m-auto">
 	<!-- Call to action to introductory sections -->
 	<section class="">
-		<h1 class="w-full my-5 pt-3 pb-5 text-5xl border-t-4 border-b-4 border-slate-500 text-center">
+		<h1
+			class="w-full my-5 pt-3 pb-5 lg:text-5xl text-3xl border-t-4 border-b-4 border-slate-500 text-center"
+		>
 			People over profits
 		</h1>
-		<div class=" grid grid-cols-2">
-			<a href="/startHere" class="w-full bg-sky-700" role="button">New? Start here!</a>
-			<a href="/startHere" class="w-full bg-green-700" role="button">Got Questions?</a>
-		</div>
+		<!-- <div class=" grid grid-cols-2"> -->
+		<!-- <a href="/startHere" class="w-full bg-sky-700" role="button">New? Start here!</a> -->
+		<a href="/startHere" class="w-full bg-green-700" role="button">Got Questions?</a>
+		<!-- </div> -->
 	</section>
 
-	<section class="grid grid-cols-3">
-		{#each cards as { question, answer }}
-			<FlipCard {question} {answer} />
-		{/each}
-	</section>
+	<Tabs />
 
-	<!-- Welcome text and intro wealth inequality video -->
-	<section class="">
-		<div>
-			<p class="my-5 indent-10">
-				Welcome to a platform for learning how the world works, what can be done about it, and how
-				you can get involved. Check out the intro video to learn why we bothered creating this
-				space.
-			</p>
-			<p class="indent-10">
-				If you're not sure where to start, check out the guided learning path for a structured
-				school-like learning experience, or ask some common questions in the questions section.
-				Either way, browse around and learn how we can change the world for the better.
-			</p>
-		</div>
-		<!-- svelte-ignore a11y-media-has-caption -->
-		<div class="w-full flex justify-center my-10">
-			<iframe
-				class="w-full aspect-video max-w-4xl rounded-lg"
-				src="https://www.youtube.com/embed/uWSxzjyMNpU"
-				title="YouTube video player"
-				frameborder="0"
-				allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-				allowfullscreen
-			/>
-		</div>
-	</section>
-
-	<!-- Interactive wealth inequality -->
-	<section>
-		<hgroup>
-			<h2 class="text-3xl mb-4">
-				Would taxing the richest handful of billionaires solve anything?
-			</h2>
-			<h3 class="indent-10">
-				It's almost impossible to visualize a billion dollars, let alone several trillion dollars,
-				the net wealth of around a 2,000 or so billionaires. Scroll to the right in the below
-				graphic to better understand just how imbalanced our system is that it allows and encourages
-				around 2,000 people to keep so much of the profits while so many of the worlds problems
-				could be solved by a fraction of it.
-				<br />
-				<br />
-				Just keep scrolling, trust us! Take about 5 minutes to keep scrolling.
-			</h3>
-		</hgroup>
-		<div class="relative">
-			<figure>
-				<iframe
-					src="https://mkorostoff.github.io/1-pixel-wealth/"
-					height="600"
-					class="w-full my-10 rounded-lg"
-					title="Wealth visualized"
-				/>
-				<figcaption class="absolute right-5 bottom-10">
-					<a href="https://mkorostoff.github.io/1-pixel-wealth/"
-						>https://mkorostoff.github.io/1-pixel-wealth</a
-					>
-				</figcaption>
-			</figure>
-		</div>
-	</section>
-
-	<!-- Browse Categories -->
-	<section>
-		<hgroup>
-			<h2 class="pt-3 pb-5 text-5xl border-t-4 border-b-4 border-slate-500 text-center">
-				Browse Categories
-			</h2>
-			<h3 class="px-10 mt-3">
-				Browse any content you'd like to educate yourself or learn how to get involved. We recommend
-				starting with myths & FAQs or with courses.
-			</h3>
-		</hgroup>
-
-		<!-- Grid of categories -->
-		<div class="md:grid md:grid-cols-2 md:gap-16">
-			<!-- Courses -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-education.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@raddfilms?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Redd F</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/9o8YdYGTT64?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Courses</h4>
-				<p class="text-center">
-					Browse courses to learn new topics or help guide your learning journey.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Questions & FAQ -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-questions.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@towfiqu999999?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Towfiqu barbhuiya</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/9o8YdYGTT64?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Myths & FAQs</h4>
-				<p class="text-center">
-					Search through all of the myths and frequently asked questions about socialism,
-					capitalism, and all the other isms.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Digital Tools -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-digital-tools.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@marvelous?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Marvin Meyer</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/SYTO3xs06fU?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Digital Tools</h4>
-				<p class="text-center">
-					Looking for help forming a union? Want help fundraising for a campaign? Want to develop
-					marketing material for your campaign? Check out our library of digital tools to help with
-					organizing, campaigning, and activism.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Research & Resources -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-research.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@homajob?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Scott Graham</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/5fNmWej4tAA?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Research & Resources</h4>
-				<p class="text-center">
-					Looking for help forming a union? Want help fundraising for a campaign? Want to develop
-					marketing material for your campaign?
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Unions -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-unions.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@mannyb?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Manny Becerra</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/GTPK-Wt1qoY?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Unions</h4>
-				<p class="text-center">
-					Learn why unions are so important, what your rights are, and how to use unions to improve
-					your life.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Policies -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-policies.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@glenncarstenspeters?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Glenn Carstens-Peters</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/RLw-UC03Gwc?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Policies</h4>
-				<p class="text-center">
-					Learn what to fight for and how to determine a good political leader from a bad one.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Get Involved -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-activism.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@vidarnm?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Vidar Nordli-Mathisen</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/cR03g2ud84g?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Get Involved</h4>
-				<p class="text-center">
-					Learn how you can run for a political position, volunteer, fund raise, organize, and
-					otherwise contribute to the movement.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Readings $ Theory -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-readings.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@vonshnauzer?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Egor Myznik</a
-						>
-						on
-						<a
-							href="https://unsplash.com/photos/1dx1gvuT45o?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Readings & Theory</h4>
-				<p class="text-center">
-					Dig deeper into theory and discovery historic events, impactful authors, and world-view
-					altering perspectives.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Readings & Theory -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-entertainment.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/@cdx2?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">C D-X</a
-						>
-						on
-						<a
-							href="https://unsplash.com/images/things/music?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Entertainment</h4>
-				<p class="text-center">
-					Find music, movie, website, video, podcast, game, and more suggestions related to
-					progressivism and the socialist movement.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-
-			<!-- Health & Lifestyle -->
-			<div class="mb-10">
-				<figure class="relative">
-					<img
-						class="object-cover w-full h-[300px] rounded-xl"
-						src="/images/category-health.jpg"
-						alt="education category"
-					/>
-					<figcaption
-						class="absolute right-0 bottom-0 z-10 text-sky-100 bg-slate-900 px-2 rounded-tl-full pl-5 bg-opacity-90 text-xs"
-					>
-						Photo by <a
-							href="https://unsplash.com/es/@esdesignisms?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-sky-300 underline">Emma Simpson</a
-						>
-						on
-						<a
-							href="https://unsplash.com/images/things/health?utm_source=unsplash&utm_medium=referral&utm_content=creditCopyText"
-							class="text-lime-300 underline">Unsplash</a
-						>
-					</figcaption>
-				</figure>
-				<h4 class="text-center uppercase mt-5 text-5xl mb-10">Health & Lifestyle</h4>
-				<p class="text-center">
-					Discovery bits of information and tips on lifestyle management and healthy living.
-				</p>
-				<a href="/" class="w-full bg-neutral-900 border-neutral-500 mt-5" role="button">Explore</a>
-			</div>
-		</div>
-	</section>
+	<Categories />
 
 	<!-- Feedback and donation -->
 	<section>
